+++ src/components/main/main-promotion-section.tsx
