+++ src/components/main/main-tip-section.tsx
